requirements.txt
python-dotenv
streamlit
google-generativeai
Pillow 
openai
PyMuPDF
numpy
pandas
flask